requirements.txt
appdirs==1.4.4
appnope==0.1.0
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
dnspython==2.0.0
docopt==0.6.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.3
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pipenv==2020.11.15
pipreqs==0.4.10
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.7.1
pylint==2.6.0
pymongo==3.11.1
pyparsing==2.4.7
pysle==2.3.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
pyzmq==19.0.2
requests==2.25.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
SQLAlchemy==1.3.22
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
tweepy==3.9.0
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
yarg==0.1.9
